add dashboard component and page
