+++ client/src/App.js
@@ -9,7 +9,7 @@ import ApolloClient from 'apollo-boost';
  import Login from "./pages/Login";
  import Signup from "./pages/Signup";
  import Nav from "./components/Nav";
-import DashboardForm from "./pages/Dashboard";
+import DashboardForm from "./components/Dashboard";
 // import { StoreProvider } from "./utils/GlobalState";
 // import Success from "./pages/Success";
 // import OrderHistory from "./pages/OrderHistory";
